feat: extracting common components and functions
