+++ packages/pluggableWidgets/datagrid-number-filter-web/src/utils/filters.ts
@@ -1,6 +1,6 @@
 import { DefaultFilterEnum } from "../../typings/DatagridNumberFilterProps";
 import { Big } from "big.js";
-import { FilterValue } from "./provider";
+import { FilterValue } from "@mendix/piw-utils-internal";
 
 export declare type DefaultFilterValue = {
     type: DefaultFilterEnum;
