+++ packages/pluggableWidgets/datagrid-web/src/Datagrid.tsx
@@ -5,9 +5,9 @@ import { and } from "mendix/filters/builders";
 
 import { Table, TableColumn } from "./components/Table";
 import classNames from "classnames";
-import { isAvailable } from "@mendix/piw-utils-internal";
+import { FilterFunction, isAvailable } from "@mendix/piw-utils-internal";
 import { extractFilters } from "./utils/filters";
-import { FilterContext, FilterFunction } from "./components/provider";
+import { FilterContext } from "./components/provider";
 
 export default function Datagrid(props: DatagridContainerProps): ReactElement {
     const [sortParameters, setSortParameters] = useState<{ columnIndex: number; desc: boolean } | undefined>(undefined);
