+++ packages/tools/piw-utils-internal/src/functions.ts
@@ -24,3 +24,17 @@ export const parseStyle = (style = ""): { [key: string]: string } => {
         return {};
     }
 };
+
+export const debounce = <F extends (...args: any[]) => any>(func: F, waitFor: number) => {
+    let timeout: ReturnType<typeof setTimeout> | null = null;
+
+    const debounced = (...args: Parameters<F>) => {
+        if (timeout !== null) {
+            clearTimeout(timeout);
+            timeout = null;
+        }
+        timeout = setTimeout(() => func(...args), waitFor);
+    };
+
+    return debounced as (...args: Parameters<F>) => ReturnType<F>;
+};
