+++ packages/tools/piw-utils-internal/src/index.ts
@@ -3,3 +3,4 @@ export * from "./functions";
 export * from "./builders";
 export * from "./utils";
 export * from "./typings";
+export * from "./widgets";
